requirements.txt
torchtext==0.4.0
transformers==4.4.2
numpy==1.19.2
pytorch_ignite==0.4.7
ignite==1.1.0
scikit_learn==1.0.2
